refactor module structure for consistency and clarity
Moved customer, account, and transaction-related files to a more coherent `features` module structure. Updated import paths and routing definitions to reflect the new file organization. This enhances maintainability and aligns the codebase with best practices.
