+++ src/app/app.routes.ts
@@ -1,9 +1,9 @@
-import { Routes } from "@angular/router";
-import { PageNotFoundComponent } from "./shared/components/page-not-found/page-not-found.component";
-import { HomeComponent } from "./shared/components/home/home.component";
-import { authGuard } from "./core/auth/auth.guard";
-import { AuthService } from "./core/auth/services/auth.service";
-import { inject } from "@angular/core";
+import { Routes } from '@angular/router';
+import { PageNotFoundComponent } from './shared/components/page-not-found/page-not-found.component';
+import { HomeComponent } from './shared/components/home/home.component';
+import { authGuard } from './core/auth/auth.guard';
+import { AuthService } from './core/auth/services/auth.service';
+import { inject } from '@angular/core';
 
 /**
  * Routing Strategy: Lazy Loading
@@ -26,31 +26,41 @@ export const routes: Routes = [
   {
     path: 'login',
     loadComponent: () =>
-      import('./core/auth/pages/login/login.component').then((m) => m.LoginComponent),
+      import('./core/auth/pages/login/login.component').then(
+        (m) => m.LoginComponent
+      ),
     canActivate: [authGuard],
   },
   {
     path: 'sign-up',
     loadComponent: () =>
-      import('./core/auth/pages/sign-up/sign-up.component').then((m) => m.SignUpComponent),
+      import('./core/auth/pages/sign-up/sign-up.component').then(
+        (m) => m.SignUpComponent
+      ),
     canActivate: [authGuard],
   },
   {
     path: 'customer',
     loadComponent: () =>
-      import('./customers/customer/customer.component').then((m) => m.CustomerComponent),
+      import('./features/customer/pages/customer/customer.component').then(
+        (m) => m.CustomerComponent,
+      ),
     canActivate: [() => inject(AuthService).isAuthenticated],
   },
   {
     path: 'accounts',
     loadComponent: () =>
-      import('./accounts/account/account.component').then((m) => m.AccountComponent),
+      import('./features/account/pages/account/account.component').then(
+        (m) => m.AccountComponent,
+      ),
     canActivate: [() => inject(AuthService).isAuthenticated],
   },
   {
     path: 'transactions',
     loadComponent: () =>
-      import('./transactions/transaction/transaction.component').then((m) => m.TransactionComponent),
+      import('./features/transaction/pages/transaction/transaction.component').then(
+        (m) => m.TransactionComponent,
+      ),
     canActivate: [() => inject(AuthService).isAuthenticated],
   },
   {
